style(!bash): remove useless blank lines
Here are some dummy messages to fill up the content of the commit description,
which looks more in details.

And there are some dummy messages in second line, let us talk more about it and
make this line warp in the commit messages.

BREAKING CHANGE: Here are some breaking changes dummy messages.

Reviewed-by: Tom
Refs: #21, #22
Closes: #23, #35